added class definitions into async server header
